add the ability to easily add Patches to the metadata
I've found there are quite a few issues with the metadata and Pull Requests to fix them don't seem forthcoming.  So I've put in some work to be able to patch the metadata with a declarative data-structure.  This will allow me to easily add/remove patches and detect when they are no longer needed.